+++ common-lib/src/main/java/n11bootcamp_project_backend/common/constants/RabbitConstants.java
@@ -1,0 +1,9 @@
+package n11bootcamp_project_backend.common.constants;
+
+public final class RabbitConstants {
+    public static final String LOG_QUEUE = "log.queue";
+    public static final String LOG_EXCHANGE = "log.exchange";
+    public static final String LOG_ROUTING_KEY = "log.routing.key";
+
+    private RabbitConstants() {}
+}