+++ common-lib/src/main/java/n11bootcamp_project_backend/common/dto/LogMessage.java
@@ -1,0 +1,16 @@
+package n11bootcamp_project_backend.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LogMessage {
+    private String serviceName;
+    private String level;
+    private String message;
+}