+++ common-lib/src/main/java/n11bootcamp_project_backend/common/dto/OrderItemQuantityResponse.java
@@ -1,0 +1,17 @@
+package n11bootcamp_project_backend.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderItemQuantityResponse {
+    private UUID productId;
+    private Integer quantity;
+}