+++ log-service/src/main/java/n11bootcamp_project_backend/log_service/LogServiceApplication.java
@@ -3,7 +3,9 @@ package n11bootcamp_project_backend.log_service;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {
+		"n11bootcamp_project_backend"
+})
 public class LogServiceApplication {
 
 	public static void main(String[] args) {