+++ log-service/src/main/java/n11bootcamp_project_backend/log_service/config/RabbitMQConfig.java
@@ -1,12 +1,12 @@
 package n11bootcamp_project_backend.log_service.config;
 
+import n11bootcamp_project_backend.common.constants.RabbitConstants;
 import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
 import org.springframework.amqp.support.converter.MessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,32 +15,26 @@ import org.springframework.context.annotation.Configuration;
 public class RabbitMQConfig {
 
     @Bean
-    public TopicExchange loggingExchange() {
-        return new TopicExchange("logging.exchange");
+    public TopicExchange logExchange() {
+        return new TopicExchange(RabbitConstants.LOG_EXCHANGE);
     }
 
     @Bean
     public Queue logQueue() {
-        return new Queue("log.queue", true);
+        return new Queue(RabbitConstants.LOG_QUEUE, true);
     }
 
     @Bean
-    public Binding logBinding(Queue logQueue, TopicExchange loggingExchange) {
-        return BindingBuilder
-                .bind(logQueue)
-                .to(loggingExchange)
-                .with("log.event");
+    public Binding logBinding(Queue logQueue, TopicExchange logExchange) {
+        return BindingBuilder.bind(logQueue).to(logExchange).with(RabbitConstants.LOG_ROUTING_KEY);
     }
 
     @Bean
-    public MessageConverter jsonMessageConverter() {
-        return new Jackson2JsonMessageConverter();
-    }
-
-    @Bean
-    public RabbitTemplate rabbitTemplate(ConnectionFactory connectionFactory) {
-        RabbitTemplate template = new RabbitTemplate(connectionFactory);
-        template.setMessageConverter(jsonMessageConverter());
-        return template;
+    public SimpleRabbitListenerContainerFactory rabbitListenerContainerFactory(
+            ConnectionFactory connectionFactory, MessageConverter jsonMessageConverter) {
+        SimpleRabbitListenerContainerFactory factory = new SimpleRabbitListenerContainerFactory();
+        factory.setConnectionFactory(connectionFactory);
+        factory.setMessageConverter(jsonMessageConverter);
+        return factory;
     }
 }